+++ src/apollo/helpers/results.ts
@@ -1,12 +1,10 @@
 import gql from 'graphql-tag';
 
-// reset current page number to 0 (first page), orderKey and order
 export const resetResultsTable = (cache: any) => {
   const { resultsTable } = cache.readQuery({
     query: gql`
       {
         resultsTable {
-          rowsPerPage
           timeFormat
         }
       }
@@ -16,9 +14,6 @@ export const resetResultsTable = (cache: any) => {
     data: {
       resultsTable: {
         ...resultsTable,
-        page: 0,
-        orderKey: '',
-        order: 'desc',
       },
     },
   });