+++ src/components/theme.ts
@@ -2,9 +2,7 @@ import { createMuiTheme } from '@material-ui/core/styles';
 import { grey, cyan } from '@material-ui/core/colors';
 
 const theme = createMuiTheme({
-  typography: {
-    useNextVariants: true,
-  },
+  typography: {},
   palette: {
     // type: 'dark',
     primary: {