+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## master
+
+#### :rocket: New Features
+
+- Material UI update
+
+#### :bug: Bug Fix
+
+-
+
 ## v0.1.6
 
 #### :bug: Bug Fix